# packages/backend/src/generated/swagger.json
{
    "components": {
...
        "schemas": {
            "ApiErrorPayload": {
                "properties": {
                    "error": {
                        "properties": {
                            "data": {
                                "description": "Optional data containing details of the error"
                            },
                            "message": {
                                "type": "string",
                                "description": "A friendly message summarising the error"
                            },
                            "name": {
                                "type": "string",
                                "description": "Unique name for the type of error"
                            },
                            "statusCode": {
                                "type": "number",
                                "format": "integer",
                                "description": "HTTP status code"
                            }
                        },
                        "required": ["name", "statusCode"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["error"],
                        "nullable": false
                    }
                },
                "required": ["error", "status"],
                "type": "object",
                "description": "The Error object is returned from the api any time there is an error.\nThe message contains"
            },
...
            "ResourceViewDashboardItem": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Pick_DashboardBasicDetails.uuid-or-spaceUuid-or-description-or-name-or-views-or-firstViewedAt-or-pinnedListUuid-or-pinnedListOrder-or-updatedAt-or-updatedByUser-or-validationErrors_"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ResourceViewItemType.DASHBOARD"
                    }
                },
                "required": ["data", "type"],
                "type": "object"
            },
...
            "ResourceViewChartItem": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Pick_SpaceQuery.uuid-or-name-or-chartType-or-firstViewedAt-or-views-or-pinnedListUuid-or-pinnedListOrder-or-spaceUuid-or-description-or-updatedAt-or-updatedByUser-or-validationErrors_"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ResourceViewItemType.CHART"
                    }
                },
                "required": ["data", "type"],
                "type": "object"
            },
...
            "ResourceViewSpaceItem": {
                "properties": {
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Pick_Space.projectUuid-or-uuid-or-name-or-isPrivate-or-pinnedListUuid-or-pinnedListOrder-or-organizationUuid_"
                            },
                            {
                                "properties": {
                                    "chartCount": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "dashboardCount": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "accessListLength": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "access": {
                                        "items": {
                                            "type": "string"
                                        },
                                        "type": "array"
                                    }
                                },
                                "required": [
                                    "chartCount",
                                    "dashboardCount",
                                    "accessListLength",
                                    "access"
                                ],
                                "type": "object"
                            }
                        ]
                    },
                    "type": {
                        "$ref": "#/components/schemas/ResourceViewItemType.SPACE"
                    }
                },
                "required": ["data", "type"],
                "type": "object"
            },
            "PinnedItems": {
                "items": {
                    "anyOf": [
                        {
                            "$ref": "#/components/schemas/ResourceViewDashboardItem"
                        },
                        {
                            "$ref": "#/components/schemas/ResourceViewChartItem"
                        },
                        {
                            "$ref": "#/components/schemas/ResourceViewSpaceItem"
                        }
                    ]
                },
                "type": "array"
            },
            "ApiPinnedItems": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/PinnedItems"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "ResourceViewItemType": {
                "enum": ["chart", "dashboard", "space"],
                "type": "string"
            },
            "Pick_ResourceViewItem-at-data.uuid-or-pinnedListOrder_": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "pinnedListOrder": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    }
                },
                "required": ["uuid", "pinnedListOrder"],
                "type": "object",
                "description": "From T, pick a set of properties whose keys are in the union K"
            },
            "UpdatePinnedItemOrder": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Pick_ResourceViewItem-at-data.uuid-or-pinnedListOrder_"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ResourceViewItemType"
                    }
                },
                "required": ["data", "type"],
                "type": "object"
            },
...
            "Project": {
                "properties": {
                    "dbtVersion": {
                        "$ref": "#/components/schemas/SupportedDbtVersions"
                    },
                    "copiedFromProjectUuid": {
                        "type": "string"
                    },
                    "pinnedListUuid": {
                        "type": "string"
                    },
                    "warehouseConnection": {
                        "$ref": "#/components/schemas/WarehouseCredentials"
                    },
                    "dbtConnection": {
                        "$ref": "#/components/schemas/DbtProjectConfig"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ProjectType"
                    },
                    "name": {
                        "type": "string"
                    },
                    "projectUuid": {
                        "type": "string"
                    },
                    "organizationUuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "dbtVersion",
                    "dbtConnection",
                    "type",
                    "name",
                    "projectUuid",
                    "organizationUuid"
                ],
                "type": "object"
            },
            "ApiProjectResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/Project"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
...
            "FieldId": {
                "type": "string"
            },
...
            "SortField": {
                "properties": {
                    "descending": {
                        "type": "boolean"
                    },
                    "fieldId": {
                        "type": "string"
                    }
                },
                "required": ["descending", "fieldId"],
                "type": "object"
            },
...
            "TableCalculation": {
                "properties": {
                    "format": {
                        "$ref": "#/components/schemas/TableCalculationFormat"
                    },
                    "sql": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "index": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": ["sql", "displayName", "name"],
                "type": "object"
            },
...
            "AdditionalMetric": {
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/MetricType"
                    },
                    "description": {
                        "type": "string"
                    },
                    "sql": {
                        "type": "string"
                    },
                    "hidden": {
                        "type": "boolean"
                    },
                    "round": {
                        "type": "number",
                        "format": "double"
                    },
                    "compact": {
                        "$ref": "#/components/schemas/CompactOrAlias"
                    },
                    "format": {
                        "$ref": "#/components/schemas/Format"
                    },
                    "table": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "index": {
                        "type": "number",
                        "format": "double"
                    },
                    "filters": {
                        "items": {
                            "$ref": "#/components/schemas/MetricFilterRule"
                        },
                        "type": "array"
                    },
                    "baseDimensionName": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": ["type", "sql", "table", "name"],
                "type": "object",
                "additionalProperties": false
            },
            "MetricQueryResponse": {
                "properties": {
                    "additionalMetrics": {
                        "items": {
                            "$ref": "#/components/schemas/AdditionalMetric"
                        },
                        "type": "array"
                    },
                    "tableCalculations": {
                        "items": {
                            "$ref": "#/components/schemas/TableCalculation"
                        },
                        "type": "array"
                    },
                    "limit": {
                        "type": "number",
                        "format": "double"
                    },
                    "sorts": {
                        "items": {
                            "$ref": "#/components/schemas/SortField"
                        },
                        "type": "array"
                    },
                    "filters": {
                        "$ref": "#/components/schemas/Filters"
                    },
                    "metrics": {
                        "items": {
                            "$ref": "#/components/schemas/FieldId"
                        },
                        "type": "array"
                    },
                    "dimensions": {
                        "items": {
                            "$ref": "#/components/schemas/FieldId"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "tableCalculations",
                    "limit",
                    "sorts",
                    "filters",
                    "metrics",
                    "dimensions"
                ],
                "type": "object"
            },
            "ApiRunQueryResponse": {
                "properties": {
                    "results": {
                        "properties": {
                            "rows": {
                                "items": {},
                                "type": "array"
                            },
                            "metricQuery": {
                                "$ref": "#/components/schemas/MetricQueryResponse"
                            }
                        },
                        "required": ["rows", "metricQuery"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "RunQueryRequest": {
                "properties": {
                    "csvLimit": {
                        "type": "number",
                        "format": "double"
                    },
                    "additionalMetrics": {
                        "items": {
                            "$ref": "#/components/schemas/AdditionalMetric"
                        },
                        "type": "array"
                    },
                    "tableCalculations": {
                        "items": {
                            "$ref": "#/components/schemas/TableCalculation"
                        },
                        "type": "array"
                    },
                    "limit": {
                        "type": "number",
                        "format": "double"
                    },
                    "sorts": {
                        "items": {
                            "$ref": "#/components/schemas/SortField"
                        },
                        "type": "array"
                    },
                    "filters": {
                        "properties": {
                            "metrics": {},
                            "dimensions": {}
                        },
                        "type": "object"
                    },
                    "metrics": {
                        "items": {
                            "$ref": "#/components/schemas/FieldId"
                        },
                        "type": "array"
                    },
                    "dimensions": {
                        "items": {
                            "$ref": "#/components/schemas/FieldId"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "tableCalculations",
                    "limit",
                    "sorts",
                    "filters",
                    "metrics",
                    "dimensions"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "session_cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            },
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Value should be 'ApiKey <your key>'"
            }
        }
    },
    "info": {
        "title": "Lightdash API",
        "version": "0.714.0",
        "description": "Open API documentation for all public Lightdash API endpoints",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Lightdash Support",
            "email": "support@lightdash.com",
            "url": "https://docs.lightdash.com/help-and-contact/contact/contact_info/"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/projects/{projectUuid}/pinned-lists/{pinnedListUuid}/items": {
            "get": {
                "operationId": "getPinnedItems",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiPinnedItems"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get pinned items",
                "tags": ["Content"],
                "security": [],
                "parameters": [
                    {
                        "description": "project uuid",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "the list uuid for the pinned items",
                        "in": "path",
                        "name": "pinnedListUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/projects/{projectUuid}/pinned-lists/{pinnedListUuid}/items/order": {
            "patch": {
                "operationId": "updatePinnedItemsOrder",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiPinnedItems"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Update pinned items order",
                "tags": ["Content"],
                "security": [],
                "parameters": [
                    {
                        "description": "project uuid",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "the list uuid for the pinned items",
                        "in": "path",
                        "name": "pinnedListUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "the new order of the pinned items",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "$ref": "#/components/schemas/UpdatePinnedItemOrder"
                                },
                                "type": "array",
                                "description": "the new order of the pinned items"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{projectUuid}": {
            "get": {
                "operationId": "GetProject",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProjectResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get a project of an organiztion",
                "tags": ["Projects"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/v1/projects/{projectUuid}/explores/{exploreId}/runUnderlyingDataQuery": {
            "post": {
                "operationId": "postRunUnderlyingDataQuery",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRunQueryResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Run a query for underlying data results",
                "tags": ["Exploring"],
                "security": [],
                "parameters": [
                    {
                        "description": "The uuid of the project",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "table name",
                        "in": "path",
                        "name": "exploreId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "metricQuery for the chart to run",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RunQueryRequest",
                                "description": "metricQuery for the chart to run"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{projectUuid}/explores/{exploreId}/runQuery": {
            "post": {
                "operationId": "RunMetricQuery",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRunQueryResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Run a query for explore",
                "tags": ["Exploring"],
                "security": [],
                "parameters": [
                    {
                        "description": "The uuid of the project",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "table name",
                        "in": "path",
                        "name": "exploreId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "metricQuery for the chart to run",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RunQueryRequest",
                                "description": "metricQuery for the chart to run"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}